Use SizeInBytes

+++ src/main/java/org/apache/hadoop/ozone/benchmark/AsyncWriter.java
@@ -41,7 +41,7 @@ public class AsyncWriter extends Writer {
   }
 
   @Override
-  public AsyncWriter init(int fileSize, ReplicationConfig replication, OzoneBucket bucket) throws IOException {
+  public AsyncWriter init(long fileSize, ReplicationConfig replication, OzoneBucket bucket) throws IOException {
     for (int i = 0; i < getPaths().size(); i++) {
       outs.add(bucket.createKey( "ozonekey_" + i, fileSize, replication, new HashMap<>()));
     }
@@ -67,7 +67,7 @@ public class AsyncWriter extends Writer {
   }
 
   @Override
-  public Map<String, CompletableFuture<Boolean>> write(int fileSize, int chunkSize, ExecutorService executor) {
+  public Map<String, CompletableFuture<Boolean>> write(long fileSize, int chunkSize, ExecutorService executor) {
     final Map<String, CompletableFuture<Boolean>> fileMap = new HashMap<>();
     for(int i = 0; i < getPaths().size(); i ++) {
       final String path = getPath(i);
